top 5 trips

+++ services/commentService.js
@@ -5,19 +5,66 @@ async function create(comment) {
 }
 
 
-async function getCommentsByTripId(tripId){
-   return Comment.find({'_tripId': `${tripId}`})
+async function getCommentsByTripId(tripId) {
+    return Comment.find({ '_tripId': `${tripId}` })
 }
 
 
-async function deleteCommentById(id){
-    return Comment.findByIdAndDelete(id)
+async function deleteCommentById(id) {
+
+    const existing = Comment.findById(id)
+
+    if (existing) {
+        return Comment.findByIdAndDelete(id)
+    } else {
+        throw new Error('Comment not found')
+
+
+    }
+
 }
+
+
+
+async function deleteCommentByOwnerId(id) {
+
+    const existing = Comment.find({ '_tripId': `${id}` })
+    if (existing) {
+        return Comment.deleteMany({ '_tripId': `${id}` })
+    } else {
+        throw new Error('Comment not found')
+
+
+    }
+
+}
+
+async function getAllMyComments(ownerId) {
+
+    const existing = Comment.find({ '_ownerId': `${ownerId}` })
+
+    if (existing) {
+        return existing
+    } else {
+        throw new Error('Comments not found')
+    }
+
+}
+
+
+
 
 
 async function getCommentById(id) {
-    console.log(id,'-----')
-    return Comment.findById(id)
+    const existing = Comment.findById(id)
+
+    if (existing) {
+        return existing
+    } else {
+        throw new Error('Comment not found')
+
+
+    }
 }
 
 async function updateCommentById(id, comment) {
@@ -26,15 +73,17 @@ async function updateCommentById(id, comment) {
     const existing = await Comment.findById(id)
 
     existing.comment = comment.comment;
-    
+
     return existing.save()
 }
 
 
-module.exports={
+module.exports = {
     create,
     getCommentsByTripId,
     deleteCommentById,
     getCommentById,
-    updateCommentById
+    updateCommentById,
+    deleteCommentByOwnerId,
+    getAllMyComments
 }
